requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# LLM and AI
llama-index>=0.9.0
transformers>=4.35.0
torch>=2.0.0
tokenizers>=0.15.0

# File processing
openpyxl>=3.1.0
chardet>=5.2.0
pyarrow>=14.0.0

# Data validation and configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional: OpenAI integration
openai>=1.3.0

# Optional: HuggingFace datasets
datasets>=2.15.0

# Optional: SQL support
duckdb>=0.9.0
sqlalchemy>=2.0.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
pyyaml>=6.0.0